Se agregan permisos al sistema

+++ app/controllers/notes_controller.rb
@@ -6,33 +6,40 @@ class NotesController < ApplicationController
 
   def index
     @notes = Note.paginate(page: params[:page], per_page: 10).order('id DESC')
+    authorize @notes
   end
 
   def show
+    authorize @note
     add_breadcrumb "#{@note.title}", note_path
   end
 
   def new
     add_breadcrumb 'New', new_note_path
     @note = Note.new
+    authorize @note
   end
 
   def edit
+    authorize @note
     add_breadcrumb 'Edit', edit_note_path
   end
 
   def create
     @note = Note.new(note_params)
+    authorize @note
     flash[:notice] = 'Note was successfully created.' if @note.save
     respond_with(@note)
   end
 
   def update
+    authorize @note
     flash[:notice] = 'Note was successfully updated.' if @note.update(note_params)
     respond_with(@note)
   end
 
   def destroy
+    authorize @note
     redirect_to notes_url, notice: 'Note was successfully destroyed.' if @note.destroy
   end
