+++ app/controllers/roles_controller.rb
@@ -6,33 +6,40 @@ class RolesController < ApplicationController
 
   def index
     @roles = Role.paginate(page: params[:page], per_page: 10).order('id DESC')
+    authorize @roles
   end
 
   def show
+    authorize @role
     add_breadcrumb "#{@role.name}", role_path
   end
 
   def new
     add_breadcrumb 'New', new_role_path
     @role = Role.new
+    authorize @role
   end
 
   def edit
+    authorize @role
     add_breadcrumb 'Edit', edit_role_path
   end
 
   def create
     @role = Role.new(role_params)
+    authorize @role
     flash[:notice] = 'Role was successfully created.' if @role.save
     respond_with(@role)
   end
 
   def update
+    authorize @role
     flash[:notice] = 'Role was successfully updated.' if @role.update(role_params)
     respond_with(@role)
   end
 
   def destroy
+    authorize @role
     if @role.destroy
       redirect_to roles_url, notice: 'Role was successfully destroyed.'
     else
